+++ apps/web/app/components/Gallery/types.ts
@@ -1,5 +1,13 @@
-import type { ImageGalleryContent } from '~/components/blocks/ImageGallery/types';
 import type { ImagesData } from '@plentymarkets/shop-api';
+
+export type ImageGalleryContent = {
+  index?: number;
+  thumbnails: {
+    showThumbnails?: boolean;
+    thumbnailType?: string;
+    enableHoverZoom?: boolean;
+  };
+};
 
 export type GalleryProps = {
   configuration?: ImageGalleryContent;
