+++ apps/web/app/composables/useCategoryTemplate/useCategoryTemplate.ts
@@ -10,15 +10,13 @@ import type { Block } from '@plentymarkets/shop-api';
 import homepageTemplateDataDe from './homepageTemplateDataDe.json';
 import homepageTemplateDataEn from './homepageTemplateDataEn.json';
 import categoryTemplateData from './categoryTemplateData.json';
-import productTemplateDataEn from './productTemplateDataEn.json';
-import productTemplateDataDe from './productTemplateDataDe.json';
+import productTemplateData from './productTemplateData.json';
 import { migrateImageContent } from '~/utils/migrate-image-content';
 
 const useLocaleSpecificHomepageTemplate = (locale: string) =>
   locale === 'de' ? (homepageTemplateDataDe as Block[]) : (homepageTemplateDataEn as Block[]);
 
-const useLocaleSpecificProductTemplate = (locale: string) =>
-  locale === 'de' ? (productTemplateDataDe as Block[]) : (productTemplateDataEn as Block[]);
+const useProductTemplateData = () => productTemplateData as Block[];
 
 const useCategoryTemplateData = () => categoryTemplateData as Block[];
 
@@ -82,7 +80,7 @@ export const useCategoryTemplate: UseCategoryTemplateReturn = (blocks?: string) 
     }
 
     if (!fetchedBlocks.length && type === 'product') {
-      fetchedBlocks = useLocaleSpecificProductTemplate($i18n.locale.value);
+      fetchedBlocks = useProductTemplateData();
     }
 
     if (Array.isArray(fetchedBlocks)) {
